backend_server/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.5.0
certifi==2023.11.17
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dj-user-login-history==1.0.5
Django==4.2
django-admin-honeypot==1.1.0
django-admin-honeypot-updated-2021==1.2.0
django-allauth==0.62.1
django-cors-headers==4.3.1
django-environ==0.12.0
django-ipware==7.0.1
django-jazzmin==3.0.0
django-widget-tweaks==1.5.0
django_login_history==0.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
i==2.1.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
pillow==11.0.0
pipenv==2023.11.15
platformdirs==4.1.0
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.17.2
PyJWT==2.10.1
PyMySQL==1.1.0
pyparsing==3.2.1
python-ipware==3.0.0
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==69.0.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.25.0
Werkzeug==3.0.1
whitenoise==6.8.2
